# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.CreateSiteRequest:
    properties:
      name:
        type: string
      url:
        type: string
    type: object
  domain.HealthResponse:
    properties:
      dependencies:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
      timestamp:
        type: string
      uptime:
        type: string
      version:
        type: string
    type: object
  domain.PingResponse:
    properties:
      message:
        type: string
    type: object
  domain.Site:
    properties:
      id:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
  domain.SiteStatus:
    properties:
      error:
        type: string
      last_check_time:
        type: string
      response_code:
        type: integer
      response_time:
        type: integer
      status:
        description: ok, pending, error
        type: string
      url:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для мониторинга доступности сайтов
  title: Go Service Monitor API
  version: 1.0.0
paths:
  /health:
    get:
      description: Возвращает текущий статус сервиса, аптайи, версию и статус зависимостей
        (пока пусто)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.HealthResponse'
      summary: Проверка доступности (здоровья) сервиса
      tags:
      - system
  /ping:
    get:
      description: Быстрая проверка доступности сервиса
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PingResponse'
      summary: Пинг-понг
      tags:
      - system
  /sites:
    get:
      description: Получить список всех сайтов, находящихся на мониторинге
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Site'
            type: array
      summary: Список сайтов
      tags:
      - sites
    post:
      consumes:
      - application/json
      description: Добавляет новый сайт для проверки
      parameters:
      - description: Данные сайта
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateSiteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Site'
        "400":
          description: invalid json body или ошибка валидации
          schema:
            type: string
        "409":
          description: url already exists
          schema:
            type: string
      summary: Добавить сайт
      tags:
      - sites
  /sites/{id}:
    delete:
      description: Удаляет сайт из мониторинга по его UUID
      parameters:
      - description: UUID сайта
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id format
          schema:
            type: string
        "404":
          description: site not found
          schema:
            type: string
      summary: Удалить сайт
      tags:
      - sites
  /sites/{id}/status:
    get:
      description: Возвращает результаты последней проверки сайта
      parameters:
      - description: id сайта
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SiteStatus'
        "400":
          description: invalid id format
          schema:
            type: string
        "404":
          description: site not found
          schema:
            type: string
      summary: Статус сайта
      tags:
      - sites
swagger: "2.0"